requirements.txt
alembic==1.8.1
Authlib==1.3.0
blinker==1.5
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==42.0.5
Flask==2.2.2
Flask-Dance==6.1.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
oauthlib==3.2.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
urllib3==1.26.12
URLObject==2.4.3
Werkzeug==2.2.2
zipp==3.8.1
